backend/requirements.txt
fastapi==0.109.0
uvicorn==0.27.0
python-dotenv==1.0.0
httpx==0.25.2
google-generativeai==0.3.0
mangum==0.17.0